refactor: moved the guide views to the guides dir
